Modify title to be a link to the home page
On most websites, the title of the website is a link
to the home page.
Modify the default layout such as the title in the header
is now a link to the home page. I had to override the default
css to keep the dark grey color for the title.

Signed-off-by: Francois Berder <fberder@outlook.fr>